requirements.txt
# Core
Django==5.0.0
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.1

# Database
psycopg2-binary==2.9.9  # PostgreSQL adapter

# Environment & Configuration
django-environ==1.0.0

# API Documentation
drf-yasg==1.21.7
django-filter==23.5

# Development (optional)
# pytest==8.0.0
# pytest-django==4.7.0
# factory-boy==3.3.0
# python-dotenv==1.0.0
# pytest==8.0.0
# pytest-django==4.7.0
# pytest-cov==4.1.0
# coverage==7.2.2
# factory-boy==3.3.0
